+++ .gitignore
@@ -1,4 +1,4 @@
-reports/
+/reports/
 logs/
 
 # Byte-compiled / optimized / DLL files
