+++ lib/connection/__init__.py
@@ -1,3 +1,3 @@
-from .RequestException import *  # noqa: F401
-from .Requester import *  # noqa: F401
-from .Response import *  # noqa: F401
+from .request_exception import *  # noqa: F401
+from .requester import *  # noqa: F401
+from .response import *  # noqa: F401
