+++ lib/output/__init__.py
@@ -1,2 +1,2 @@
-from .CLIOutput import *  # noqa: F401
-from .PrintOutput import *  # noqa: F401
+from .cli_output import *  # noqa: F401
+from .print_output import *  # noqa: F401
