+++ lib/output/print_output.py
@@ -18,35 +18,33 @@
 
 import sys
 import threading
-import time
 
-from lib.utils.FileUtils import *
-from lib.utils.TerminalSize import get_terminal_size
+from lib.utils.file_utils import *
 from thirdparty.colorama import *
 
 if sys.platform in ["win32", "msys"]:
     from thirdparty.colorama.win32 import *
 
 
-class CLIOutput(object):
+class PrintOutput(object):
     def __init__(self):
         init()
-        self.lastLength = 0
-        self.lastOutput = ""
-        self.lastInLine = False
         self.mutex = threading.Lock()
         self.blacklists = {}
+        self.mutexCheckedPaths = threading.Lock()
         self.basePath = None
         self.errors = 0
+
+    def header(self, text):
+        pass
 
     def inLine(self, string):
         self.erase()
         sys.stdout.write(string)
         sys.stdout.flush()
-        self.lastInLine = True
 
     def erase(self):
-        if sys.platform in ["win32", "msys"]:
+        if sys.platform in ["win32", "cygwin", "msys"]:
             csbi = GetConsoleScreenBufferInfo()
             line = "\b" * int(csbi.dwCursorPosition.X)
             sys.stdout.write(line)
@@ -61,20 +59,7 @@ class CLIOutput(object):
             sys.stdout.write("\033[0G")
 
     def newLine(self, string):
-        if self.lastInLine is True:
-            self.erase()
-
-        if sys.platform in ["win32", "msys"]:
-            sys.stdout.write(string)
-            sys.stdout.flush()
-            sys.stdout.write("\n")
-            sys.stdout.flush()
-
-        else:
-            sys.stdout.write(string + "\n")
-
-        sys.stdout.flush()
-        self.lastInLine = False
+        sys.stdout.write(string + "\n")
         sys.stdout.flush()
 
     def statusReport(self, path, response, full_url, addedToQueue):
@@ -97,22 +82,16 @@ class CLIOutput(object):
                 contentLength = FileUtils.sizeHuman(size)
 
             if not self.basePath:
-                showPath = "/" + path
+                showPath = urljoin("/", path)
 
             else:
                 if not self.basePath.startswith("/"):
                     self.basePath = "/" + self.basePath
 
                 showPath = self.basePath.rstrip("/") + "/" + path
-
-                if full_url:
-                    showPath = (self.target[:-1] if self.target.endswith("/") else self.target) + showPath
-
-            message = "[{0}] {1} - {2} - {3}".format(
-                time.strftime("%H:%M:%S"),
-                status,
-                contentLength.rjust(6, " "),
-                showPath,
+                showPath = (self.target[:-1] if self.target.endswith("/") else self.target) + showPath
+            message = "{0} - {1} - {2}".format(
+                status, contentLength.rjust(6, " "), showPath
             )
 
             if status == 200:
@@ -131,7 +110,7 @@ class CLIOutput(object):
                 message = Fore.RED + message + Style.RESET_ALL
 
             # Check if redirect
-            elif status in [301, 302, 303, 307, 308] and "location" in [
+            elif status in [301, 302, 307] and "location" in [
                 h.lower() for h in response.headers
             ]:
                 message = Fore.CYAN + message + Style.RESET_ALL
@@ -142,49 +121,17 @@ class CLIOutput(object):
 
             self.newLine(message)
 
-    @staticmethod
-    def percentage(x, y):
-        return float(x) / float(y) * 100
-
     def lastPath(self, path, index, length, currentJob, allJobs):
-        x, y = get_terminal_size()
-
-        message = "{0:.2f}% - ".format(self.percentage(index, length))
-
-        if allJobs > 1:
-            message += "Job: {0}/{1} - ".format(currentJob, allJobs)
-
-        if self.errors > 0:
-            message += "Errors: {0} - ".format(self.errors)
-
-        message += "Last request to: {0}".format(path)
-
-        if len(message) >= x:
-            message = message[:x - 1]
-
-        with self.mutex:
-            self.inLine(message)
+        pass
 
     def addConnectionError(self):
         self.errors += 1
 
     def error(self, reason):
-        with self.mutex:
-            stripped = reason.strip()
-            message = "\n" if reason.startswith("\n") else ""
-            message += Style.BRIGHT + Fore.WHITE + Back.RED
-            message += stripped
-            message += Style.RESET_ALL
-            self.newLine(message)
+        pass
 
     def warning(self, reason):
-        with self.mutex:
-            message = Style.BRIGHT + Fore.YELLOW + reason + Style.RESET_ALL
-            self.newLine(message)
-
-    def header(self, text):
-        message = Style.BRIGHT + Fore.MAGENTA + text + Style.RESET_ALL
-        self.newLine(message)
+        pass
 
     def config(
         self,
@@ -197,52 +144,16 @@ class CLIOutput(object):
         recursive,
         recursion_level,
     ):
-        separator = Fore.MAGENTA + " | " + Fore.YELLOW
-
-        config = Style.BRIGHT + Fore.YELLOW
-        config += "Extensions: {0}".format(Fore.CYAN + extensions + Fore.YELLOW)
-        config += separator
-
-        config += "HTTP method: {0}".format(Fore.CYAN + method.upper() + Fore.YELLOW)
-        config += separator
-
-        if prefixes != '':
-            config += 'Prefixes: {0}'.format(Fore.CYAN + prefixes + Fore.YELLOW)
-            config += separator
-
-        if suffixes != '':
-            config += 'Suffixes: {0}'.format(Fore.CYAN + suffixes + Fore.YELLOW)
-            config += separator
-
-        config += "Threads: {0}".format(Fore.CYAN + threads + Fore.YELLOW)
-        config += separator
-        config += "Wordlist size: {0}".format(Fore.CYAN + wordlist_size + Fore.YELLOW)
-
-        if recursive is True:
-            config += separator
-            config += "Recursion level: {0}".format(
-                Fore.CYAN + recursion_level + Fore.YELLOW
-            )
-
-        config += Style.RESET_ALL
-
-        self.newLine(config)
+        pass
 
     def setTarget(self, target):
         self.target = target
 
-        config = Style.BRIGHT + Fore.YELLOW
-        config += "\nTarget: {0}\n".format(Fore.CYAN + target + Fore.YELLOW)
-        config += Style.RESET_ALL
-
-        self.newLine(config)
-
     def outputFile(self, target):
-        self.newLine("Output File: {0}\n".format(target))
+        pass
 
     def errorLogFile(self, target):
-        self.newLine("\nError Log: {0}".format(target))
+        pass
 
     def debug(self, info):
-        line = "[{0}] - {1}".format(time.strftime("%H:%M:%S"), info)
-        self.newLine(line)
+        pass
